% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{block_bootstrap_granges_within_chrom}
\alias{block_bootstrap_granges_within_chrom}
\title{Block bootstrap GRanges within chromosome}
\usage{
block_bootstrap_granges_within_chrom(x, L_b, L_s, chr)
}
\arguments{
\item{x}{the input GRanges}

\item{L_b}{the length of the blocks}

\item{L_s}{the length of the segment (chromosome)}

\item{chr}{the name of the chromosome}
}
\description{
Block bootstrap GRanges within chromosome
}
