% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BootRanges-class}
\alias{BootRanges-class}
\alias{BootRanges}
\title{BootRanges object}
\description{
This class extends the GRanges class. It is expected to have
an \code{iter} metadata column indicating the iterations of the
bootstrap, which are concatenated together.
}
