% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootRanges.R
\name{bootRanges}
\alias{bootRanges}
\title{Block bootstrap genomic ranges}
\usage{
bootRanges(
  x,
  seg = NULL,
  L_b,
  R = 1,
  deny = NULL,
  deny_option = c("drop", "trim"),
  proportion_length = TRUE,
  type = c("bootstrap", "permute"),
  within_chrom = FALSE
)
}
\arguments{
\item{x}{the input GRanges}

\item{seg}{the segmentation GRanges, with a column ("state")
indicating segmentation state (optional)}

\item{L_b}{the length of the blocks (for proportional blocks, this
is the maximal length of block)}

\item{R}{the number of bootstrap samples to generate}

\item{deny}{the GRanges of deny regions (optional)}

\item{deny_option}{whether to "drop" or "trim" bootstrap
ranges that overlap a deny region}

\item{proportion_length}{for segmented block bootstrap,
whether to use scaled block length, (scaling by the proportion
of the segmentation state out of the total genome length)}

\item{type}{the type of null generation (unsegmented only)}

\item{within_chrom}{whether to re-sample (bootstrap) ranges
across chromosomes (default) or only within chromosomes (unsegmented only)}
}
\value{
bootRanges: a GRangesList of length R with the bootstrapped ranges
}
\description{
Performs a block bootstrap (R times) optionally with respect
to a genome segmentation. Returns a bootRanges object, which
is essentially a GRangesList of length R.
}
