% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_iranges}
\alias{bootstrap_iranges}
\title{Block bootstrap IRanges}
\usage{
bootstrap_iranges(x, L_s, L_b)
}
\arguments{
\item{x}{the input ranges}

\item{L_s}{the length of the segment}

\item{L_b}{the length of the blocks
n}
}
\description{
Block bootstrap IRanges
}
