% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unseg_bootstrap.R
\name{bootstrapRanges}
\alias{bootstrapRanges}
\title{Unsegmented block bootstrap}
\usage{
bootstrapRanges(x, L_b, type = c("bootstrap", "permute"), within_chrom = FALSE)
}
\arguments{
\item{x}{the input GRanges to bootstrap}

\item{L_b}{the length of the blocks}

\item{type}{the type of null generation}

\item{within_chrom}{whether to re-sample (bootstrap) ranges
across chromosomes (default) or only within chromosomes}
}
\value{
one iteration of bootstrapped ranges
(TODO fix this)
}
\description{
Unsegmented block bootstrap
}
