% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-utils.R
\name{combnCov}
\alias{combnCov}
\alias{combnCov,character-method}
\title{Function for creating combinations of covariates}
\usage{
combnCov(x)

\S4method{combnCov}{character}(x)
}
\arguments{
\item{x}{character vector of covariates to combine}
}
\value{
returns a character vector of formulae combinations.
}
\description{
Function for creating combinations of covariates
}
\examples{

combnCov(x = c('a', 'b', 'c'))

}
