% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-Matched.R
\name{covariates}
\alias{covariates}
\alias{covariates,Matched-method}
\title{Get covariates from a Matched object}
\usage{
covariates(x, ...)

\S4method{covariates}{Matched}(x, ...)
}
\arguments{
\item{x}{Matched object.}

\item{...}{Additional arguments.}
}
\value{
A character vector of covariates
}
\description{
Get covariates from a Matched object
}
\examples{
set.seed(123)
mdf <- makeExampleMatchedDataSet(matched = TRUE)
covariates(mdf)

}
