% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-matchRanges.R
\name{focal}
\alias{focal}
\alias{focal,MDF_OR_MGR_OR_MGI-method}
\title{Get focal set from a Matched object}
\usage{
focal(x, ...)

\S4method{focal}{MDF_OR_MGR_OR_MGI}(x, ...)
}
\arguments{
\item{x}{A \code{MatchedDataFrame}, \code{MatchedGRanges},
or \code{MatchedGInteractions} object.}

\item{...}{Additional options.}
}
\description{
Get focal set from a Matched object
}
\examples{
x <- makeExampleMatchedDataSet(matched = TRUE)
focal(x)

}
