% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-Matched.R
\name{plotPropensity}
\alias{plotPropensity}
\alias{plotPropensity,Matched,character_OR_missing,character_OR_missing,character_OR_missing-method}
\title{Plotting functions for Matched objects}
\usage{
plotPropensity(
  x,
  sets = c("focal", "matched", "pool", "unmatched"),
  type = NULL,
  log = NULL,
  ...
)

\S4method{plotPropensity}{Matched,character_OR_missing,character_OR_missing,character_OR_missing}(x, sets, type, log, thresh = 12)
}
\arguments{
\item{x}{...}

\item{type}{...}

\item{...}{additional arguments}
}
\description{
Plotting functions for Matched objects
}
