% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-matchRanges.R
\name{matched}
\alias{matched}
\alias{matched,MDF_OR_MGR_OR_MGI-method}
\title{Get matched set from a Matched object}
\usage{
matched(x, ...)

\S4method{matched}{MDF_OR_MGR_OR_MGI}(x, ...)
}
\arguments{
\item{x}{A \code{MatchedDataFrame}, \code{MatchedGRanges},
or \code{MatchedGInteractions} object.}

\item{...}{Additional options.}
}
\value{
An object of the same class as \code{x} representing
the matched set.
}
\description{
Get matched set from a Matched object
}
\examples{
set.seed(123)
x <- makeExampleMatchedDataSet(matched = TRUE)
matched(x)

}
