% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-Matched.R
\name{matchedData}
\alias{matchedData}
\alias{matchedData,Matched-method}
\title{Get matched data from a Matched object}
\usage{
matchedData(x, ...)

\S4method{matchedData}{Matched}(x, ...)
}
\arguments{
\item{x}{Matched object.}

\item{...}{Additional arguments.}
}
\value{
A \code{data.table} with matched data.
}
\description{
Get matched data from a Matched object
}
\examples{
set.seed(123)
mdf <- makeExampleMatchedDataSet(matched = TRUE)
matchedData(mdf)

}
