% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-matchRanges.R
\name{focal,MDF_OR_MGR_OR_MGI-method}
\alias{focal,MDF_OR_MGR_OR_MGI-method}
\alias{pool,MDF_OR_MGR_OR_MGI-method}
\alias{matched,MDF_OR_MGR_OR_MGI-method}
\alias{unmatched,MDF_OR_MGR_OR_MGI-method}
\title{Accessor methods for matchRanges Class}
\usage{
\S4method{focal}{MDF_OR_MGR_OR_MGI}(x, ...)

\S4method{pool}{MDF_OR_MGR_OR_MGI}(x, ...)

\S4method{matched}{MDF_OR_MGR_OR_MGI}(x, ...)

\S4method{unmatched}{MDF_OR_MGR_OR_MGI}(x, ...)
}
\arguments{
\item{x}{Matched object}

\item{...}{additional arguments}
}
\description{
Functions that get data from "matchRanges" subclasses
such as MatchedDataFrame, MatchedGRanges,
and MatchedGInteractions.
}
