% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-Matched.R
\name{method}
\alias{method}
\alias{method,Matched-method}
\title{Get matching method used for Matched object}
\usage{
method(x, ...)

\S4method{method}{Matched}(x, ...)
}
\arguments{
\item{x}{Matched object.}

\item{...}{Additional arguments.}
}
\value{
A character describing the matched method
}
\description{
Get matching method used for Matched object
}
\examples{
mdf <- makeExampleMatchedDataSet(matched = TRUE)
method(mdf)

}
