% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Matched.R
\name{names.MatchedOverview}
\alias{names.MatchedOverview}
\alias{$,MatchedOverview-method}
\alias{[[,MatchedOverview,ANY,ANY-method}
\alias{show,MatchedOverview-method}
\title{Names S3 method for autocomplete}
\usage{
\method{names}{MatchedOverview}(x)

\S4method{$}{MatchedOverview}(x, name)

\S4method{[[}{MatchedOverview,ANY,ANY}(x, i)

\S4method{show}{MatchedOverview}(object)
}
\arguments{
\item{x}{A \code{MatchedOverview} object.}

\item{name}{Name of slot.}

\item{i}{A character or numeric to extract.}

\item{object}{A \code{MatchedOverview} object.}
}
\description{
Names S3 method for autocomplete

Extract \code{$} operator for MatchedOverview

Extract \code{[[} operator for MatchedOverview

Show method for \code{overview}
}
\keyword{internal}
