% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{permute_blocks_iranges}
\alias{permute_blocks_iranges}
\title{Permute blocks IRanges}
\usage{
permute_blocks_iranges(x, L_s, L_b)
}
\arguments{
\item{x}{the input ranges}

\item{L_s}{the length of the segment}

\item{L_b}{the length of the blocks}
}
\description{
Permute blocks IRanges
}
