% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plotCovariates}
\alias{plotCovariates}
\title{plotCovariates}
\usage{
plotCovariates(x, covar, sets, type, logTransform, ...)
}
\arguments{
\item{x}{an object}

\item{covar}{...}

\item{sets}{...}

\item{type}{...}

\item{logTransform}{...}

\item{...}{additional arguments}
}
\description{
plotCovariates
}
