% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-matchRanges.R
\name{pool}
\alias{pool}
\alias{pool,MDF_OR_MGR_OR_MGI-method}
\title{Get pool set from a Matched object}
\usage{
pool(x, ...)

\S4method{pool}{MDF_OR_MGR_OR_MGI}(x, ...)
}
\arguments{
\item{x}{A \code{MatchedDataFrame}, \code{MatchedGRanges},
or \code{MatchedGInteractions} object.}

\item{...}{Additional options.}
}
\description{
Get pool set from a Matched object
}
\examples{
x <- makeExampleMatchedDataSet(matched = TRUE)
pool(x)

}
