% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_bootstrap_map.R
\name{seg_bootstrap_granges_map}
\alias{seg_bootstrap_granges_map}
\title{Segmentated bootstrap GRanges, map all chromosome to one line}
\usage{
seg_bootstrap_granges_map(
  seg,
  x,
  within_chrom = TRUE,
  L_b,
  proportion_length = TRUE,
  coarse = FALSE
)
}
\arguments{
\item{seg}{segmentation GRanges a column("state") indicate segmentation states}

\item{x}{the input GRanges}

\item{L_b}{the length of the block}

\item{proportion_length}{use scaled block length or scaled number of blocks of each segmentation region}

\item{coarse}{simply check the chromosome ranges}

\item{L_c}{the length of the chromosome}

\item{L_s}{a vector of the length of each segmentation region}
}
\description{
Segmentated bootstrap GRanges, map all chromosome to one line
}
