% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_bootstrap.R
\name{seg_bootstrap_iranges}
\alias{seg_bootstrap_iranges}
\title{Segmentated bootstrap GRanges}
\usage{
seg_bootstrap_iranges(seg, x, L_c, L_s, L_b)
}
\arguments{
\item{seg}{segmentation GRanges a column("state") indicate segmentation states}

\item{x}{the input GRanges}

\item{L_c}{the length of the chromosome}

\item{L_s}{a vector of the length of each segmentation region}

\item{L_b}{the length of the block}
}
\description{
Segmentated bootstrap GRanges
}
