% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_bootstrap.R
\name{seg_bootstrap_iranges_wchr}
\alias{seg_bootstrap_iranges_wchr}
\title{segmentation Block bootstrap IRanges within chromosome}
\usage{
seg_bootstrap_iranges_wchr(
  seg,
  x,
  state,
  L_c,
  L_b,
  proportion_length = TRUE,
  coarse = FALSE
)
}
\arguments{
\item{seg}{the ranges of segmentation GRanges}

\item{x}{the ranges of the input GRanges}

\item{state}{the segmentation state}

\item{L_c}{the chromosome length}

\item{L_b}{the length of the block}

\item{proportion_length}{use scaled block length or scaled number of blocks of each segmentation region}

\item{coarse}{logical indicating if out of bound ranges will be discarded or keep the intersect region.}
}
\description{
segmentation Block bootstrap IRanges within chromosome
}
