% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_bootstrap.R
\name{segBootstrapRanges}
\alias{segBootstrapRanges}
\title{Segmentated block bootstrap}
\usage{
segBootstrapRanges(
  x,
  seg,
  L_b,
  R,
  deny = NULL,
  deny_option = c("drop", "trim"),
  proportion_length = TRUE
)
}
\arguments{
\item{x}{the input GRanges}

\item{seg}{the segmentation GRanges, with a column ("state")
indicating segmentation state}

\item{L_b}{the length of the blocks (for proportional blocks, this
is the maximal length of block)}

\item{R}{the number of bootstrap samples to generate}

\item{deny}{the GRanges of deny regions (optional)}

\item{deny_option}{whether to "drop" or "trim" bootstrap
ranges that overlap a deny region}

\item{proportion_length}{whether to use scaled block length,
(scaling by the proportion of the segmentation state
out of the total genome length)}
}
\description{
Segmentated block bootstrap
}
