% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-matchRanges.R
\name{unmatched}
\alias{unmatched}
\alias{unmatched,MDF_OR_MGR_OR_MGI-method}
\title{Get unmatched set from a Matched object}
\usage{
unmatched(x, ...)

\S4method{unmatched}{MDF_OR_MGR_OR_MGI}(x, ...)
}
\arguments{
\item{x}{A \code{MatchedDataFrame}, \code{MatchedGRanges},
or \code{MatchedGInteractions} object.}

\item{...}{Additional options.}
}
\description{
Get unmatched set from a Matched object
}
\examples{
x <- makeExampleMatchedDataSet(matched = TRUE)
unmatched(x)

}
