% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-Matched.R
\name{withReplacement}
\alias{withReplacement}
\alias{withReplacement,Matched-method}
\title{Get replace method}
\usage{
withReplacement(x, ...)

\S4method{withReplacement}{Matched}(x, ...)
}
\arguments{
\item{x}{Matched object.}

\item{...}{Additional arguments.}
}
\value{
TRUE/FALSE indicating whether matching
was done with or without replacement.
}
\description{
Determine if a Matched object was created
with or without replacement.
}
\examples{
set.seed(123)
mdf <- makeExampleMatchedDataSet(matched = TRUE)
withReplacement(mdf)

}
