% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BootRanges-class}
\alias{BootRanges-class}
\alias{BootRanges}
\title{BootRanges object}
\description{
This class extends the GRanges class. As produced by bootRanges,
it will have an \code{iter} column indicating the iterations of the
bootstrap, and a optionally a \code{blockLength} column indicating the
blockLength parameter.
}
