% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_granges}
\alias{bootstrap_granges}
\title{Bootstrap GRanges}
\usage{
bootstrap_granges(
  x,
  L_b,
  type = c("bootstrap", "permute"),
  within_chrom = TRUE
)
}
\arguments{
\item{x}{the input GRanges}

\item{L_b}{the length of the block}

\item{type}{the type of null generation}

\item{within_chrom}{whether to re-sample (bootstrap) ranges
only wwithin each chromosome, or whether to allow
cross-chromosome re-sampling}
}
\description{
Bootstrap GRanges
}
