% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-Matched.R
\name{plotCovariates}
\alias{plotCovariates}
\alias{plot,Matched,missing-method}
\alias{plotCovariates,Matched-method}
\title{Plotting functions for Matched objects}
\usage{
plotCovariates(x, covar, sets, type, logTransform, ...)

\S4method{plot}{Matched,missing}(x, type = NULL)

\S4method{plotCovariates}{Matched}(x, covar, sets, type, logTransform, ...)
}
\arguments{
\item{x}{...}

\item{covar}{...}

\item{sets}{...}

\item{type}{...}

\item{logTransform}{...}

\item{...}{additional arguments}
}
\description{
Plotting functions for Matched objects
}
