% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-Matched.R
\name{plotPropensity}
\alias{plotPropensity}
\alias{plotPropensity,Matched-method}
\title{Plotting functions for Matched objects}
\usage{
plotPropensity(x, ...)

\S4method{plotPropensity}{Matched}(x, type = NULL)
}
\arguments{
\item{x}{...}

\item{...}{additional arguments}

\item{type}{...}
}
\description{
Plotting functions for Matched objects
}
