% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/methods-Matched.R
\docType{class}
\name{Matched-class}
\alias{Matched-class}
\alias{Matched}
\alias{matched}
\alias{matchedData,Matched-method}
\alias{covariates,Matched-method}
\alias{indices,Matched-method}
\alias{overview,Matched-method}
\title{Matched}
\usage{
matched(x, ...)

\S4method{matchedData}{Matched}(x, ...)

\S4method{covariates}{Matched}(x, ...)

\S4method{indices}{Matched}(x, group = "matched")

\S4method{overview}{Matched}(x)
}
\arguments{
\item{x}{an object}

\item{...}{additional arguments}

\item{group}{a character string describing from which group to extract indices.
can be one of 'focal', 'matched', 'pool', or 'unmatched'.}
}
\description{
Functions that get data from Matched subclasses
such as Matched, MatchedDataFrame, MatchedGRanges,
and MatchedGInteractions.
}
