% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-Matched.R
\docType{class}
\name{Matched-class}
\alias{Matched-class}
\alias{Matched}
\alias{matchedData,Matched-method}
\alias{covariates,Matched-method}
\alias{indices,Matched-method}
\alias{overview,Matched-method}
\alias{plot,Matched,missing-method}
\alias{plotCovariates,Matched-method}
\title{Accessor methods for Matched Class}
\usage{
\S4method{matchedData}{Matched}(x, ...)

\S4method{covariates}{Matched}(x, ...)

\S4method{indices}{Matched}(x, group = "matched")

\S4method{overview}{Matched}(x)

\S4method{plot}{Matched,missing}(x, type)

\S4method{plotCovariates}{Matched}(x, covar, sets, type, logTransform, ...)
}
\arguments{
\item{x}{Matched object}

\item{...}{additional arguments}

\item{group}{a character string describing from which group to extract indices.
can be one of 'focal', 'matched', 'pool', or 'unmatched'.}
}
\description{
Functions that get data from Matched subclasses
such as Matched, MatchedDataFrame, MatchedGRanges,
and MatchedGInteractions.
}
