% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-matchRanges.R
\name{matchRanges}
\alias{matchRanges}
\alias{matchRanges,DF_OR_df_OR_dt,DF_OR_df_OR_dt,formula-method}
\alias{matchRanges,GRanges,GRanges,formula-method}
\alias{matchRanges,GInteractions,GInteractions,formula-method}
\alias{focal,MDF_OR_MGR_OR_MGI-method}
\alias{pool,MDF_OR_MGR_OR_MGI-method}
\alias{matched,MDF_OR_MGR_OR_MGI-method}
\alias{unmatched,MDF_OR_MGR_OR_MGI-method}
\title{matchRanges}
\usage{
matchRanges(focal, pool, covar, ...)

\S4method{matchRanges}{DF_OR_df_OR_dt,DF_OR_df_OR_dt,formula}(focal, pool, covar)

\S4method{matchRanges}{GRanges,GRanges,formula}(focal, pool, covar)

\S4method{matchRanges}{GInteractions,GInteractions,formula}(focal, pool, covar)

\S4method{focal}{MDF_OR_MGR_OR_MGI}(x, ...)

\S4method{pool}{MDF_OR_MGR_OR_MGI}(x, ...)

\S4method{matched}{MDF_OR_MGR_OR_MGI}(x, ...)

\S4method{unmatched}{MDF_OR_MGR_OR_MGI}(x, ...)
}
\arguments{
\item{focal}{a DataFrame, GRanges, or GInteractions object containing
the focal data to match.}

\item{pool}{a DataFrame, GRanges, or GInteractions object containing
the pool from which to select matches.}

\item{covar}{a rhs formula with covariates on which to match.}

\item{...}{additional arguments}

\item{x}{Matched object}
}
\value{
a covariate-matched control set of data
}
\description{
A function that generates a covariate-matched control
set of the same datatype.

Functions that get data from "matchRanges" subclasses
such as MatchedDataFrame, MatchedGRanges,
and MatchedGInteractions.
}
