% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-matchRanges.R
\name{matchRanges}
\alias{matchRanges}
\alias{matchRanges,GRanges,GRanges,formula,character,logical-method}
\alias{matchRanges,GInteractions,GInteractions,formula,character,logical-method}
\title{matchRanges}
\usage{
matchRanges(focal, pool, covar, method, replace, ...)

\S4method{matchRanges}{DF_OR_df_OR_dt,DF_OR_df_OR_dt,formula,character,logical}(focal, pool, covar, method, replace)

\S4method{matchRanges}{GRanges,GRanges,formula,character,logical}(focal, pool, covar, method, replace)

\S4method{matchRanges}{GInteractions,GInteractions,formula,character,logical}(focal, pool, covar, method, replace)
}
\arguments{
\item{focal}{a DataFrame, GRanges, or GInteractions object containing
the focal data to match.}

\item{pool}{a DataFrame, GRanges, or GInteractions object containing
the pool from which to select matches.}

\item{covar}{a rhs formula with covariates on which to match.}

\item{method}{character describing which matching method to use.
supported options are either 'nearest' or 'rejection'.}

\item{replace}{logical describing whether to select matches with or without
replacement.}

\item{...}{additional arguments}
}
\value{
a covariate-matched control set of data
}
\description{
A function that generates a covariate-matched control
set of the same datatype.
}
