% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchRanges.R
\name{matchRanges}
\alias{matchRanges}
\title{Generate a set of matched control GRanges}
\usage{
matchRanges(x, univ, covar, method = "nearest")
}
\arguments{
\item{x}{GRanges object containing focal ranges to match}

\item{univ}{GRanges object containing the universe of ranges from which to select}

\item{covar}{Character vector of covariates to match on}

\item{method}{Character string passed to MatchIt used to calculate matches}
}
\value{
Returns a GRanges object of matched controls
}
\description{
Given an input of focal GRanges, universe GRanges and covariates, this function will
generate a set of GRanges selected from the universe that are matched on the covariates
to the focal ranges. This function wraps the covariate matching function from the
package 'MatchIt'.
}
\examples{

matchRanges(x = up_ov, univ = ctl_ov, covar = c("peakStrength", "gc"))


}
