% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootHelper.R
\name{oneRegionSegment}
\alias{oneRegionSegment}
\title{Segmentation based on one region}
\usage{
oneRegionSegment(x, seqlength)
}
\arguments{
\item{x}{a single region as GRanges object}
}
\value{
a segmentation (GRanges object) with the region of
interest designated as state 2, and the rest of the chromosome
as state 1.
}
\description{
This function makes a segmentation (GRanges) based on one
region of one chromosome (seqnames).
}
