% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSegment.R
\name{plotSegment}
\alias{plotSegment}
\title{Genomic segmentation based on gene density}
\usage{
plotSegment(
  x,
  seg,
  L_s = 1e+06,
  deny,
  type = c("ranges", "boxplot"),
  region = NULL
)
}
\arguments{
\item{x}{the input gene GRanges}

\item{seg}{the segmentation GRanges returned by \code{segmentDensity} function}

\item{L_s}{segment length}

\item{deny}{GRanges of deny region}

\item{type}{the type of plot returned. Either a segmentation plot
included ranges information or a box plot. Default is both plots are displayed.
The y axis \code{"density"} represent square root of overlap counts within segment length.}

\item{region}{GRanges of stricted region that want to be plotted.}
}
\description{
Genomic segmentation based on gene density
}
