% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSegment.R
\name{plotSegment}
\alias{plotSegment}
\title{Genomic segmentation based on gene density}
\usage{
plotSegment(
  seg,
  exclude,
  type = c("ranges", "barplot", "boxplot"),
  region = NULL
)
}
\arguments{
\item{seg}{the segmentation GRanges returned by \code{segmentDensity} function}

\item{exclude}{GRanges of excluded region}

\item{type}{the type of plot returned. Choices are segmentation plot
included ranges information, barplot showing segmentation states' distribution across chromosome,
or a box plot indicating average density within each states.
Default is all plots are displayed.
The y axis \code{"density"} represent square root of overlap counts within segment length.}

\item{region}{GRanges of stricted region that want to be plotted.}
}
\value{
A \code{ggplot} set by \code{type} argument
}
\description{
Genomic segmentation based on gene density
}
