% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_bootstrap.R
\name{seg_bootstrap_iranges}
\alias{seg_bootstrap_iranges}
\title{segmentation Block bootstrap IRanges span the whole genome.}
\usage{
seg_bootstrap_iranges(
  seg,
  x,
  state,
  L_b,
  chrnames,
  chrom_lens,
  proportion_length = TRUE
)
}
\arguments{
\item{seg}{the ranges of segmentation GRanges}

\item{x}{the input GRanges with a column ("key.x") indicate the index of gene}

\item{state}{the segmentation state}

\item{L_b}{the length of the block}

\item{chrnames}{the seqnames of the input GRanges}

\item{chrom_lens}{all chromosome length}

\item{proportion_length}{use scaled block length or scaled number of blocks of each segmentation region}

\item{coarse}{logical indicating if out of bound ranges will be discarded or keep the intersect region.}
}
\description{
segmentation Block bootstrap IRanges span the whole genome.
}
