% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_bootstrap.R
\name{segBootstrapRanges}
\alias{segBootstrapRanges}
\title{Segmentated block bootstrap}
\usage{
segBootstrapRanges(
  x,
  seg,
  L_b,
  R,
  deny,
  deny_option = c("toss", "trim"),
  within_chrom = FALSE,
  proportion_length = TRUE,
  ncores = NULL
)
}
\arguments{
\item{x}{the input GRanges}

\item{seg}{segmentation GRanges a column("state") indicate segmentation states}

\item{L_b}{the length of the block}

\item{R}{the number of bootstrap samples to generate}

\item{deny}{GRanges of deny region}

\item{deny_option}{Indicate whether toss or trim the overlaps between deny and
bootstrap ranges.}

\item{within_chrom}{whether to perform bootstrapping within chromosome (default FALSE)}

\item{proportion_length}{use scaled block length or scaled number of blocks of each segmentation region}

\item{ncores}{A cluster object created by \code{\link[parallel]{makeCluster}}.
Or an integer to indicate number of child-processes
(integer values are ignored on Windows) for parallel evaluations}
}
\description{
Segmentated block bootstrap
}
