% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_bootstrap.R
\name{segBootstrapRanges}
\alias{segBootstrapRanges}
\title{Segmentated bootstrap GRanges}
\usage{
segBootstrapRanges(
  seg,
  x,
  L_b,
  R,
  within_chrom = FALSE,
  proportion_length = TRUE,
  ncores = NULL
)
}
\arguments{
\item{seg}{segmentation GRanges a column("state") indicate segmentation states}

\item{x}{the input GRanges}

\item{L_b}{the length of the block}

\item{R}{the number time of bootstrap}

\item{proportion_length}{use scaled block length or scaled number of blocks of each segmentation region}

\item{ncores}{A cluster object created by \code{\link[parallel]{makeCluster}}.
Or an integer to indicate number of child-processes
(integer values are ignored on Windows) for parallel evaluations}

\item{L_c}{the length of the chromosome}

\item{L_s}{a vector of the length of each segmentation region}
}
\description{
Segmentated bootstrap GRanges
}
