% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_density.R
\name{segment_density}
\alias{segment_density}
\title{Segmentation based on gene density}
\usage{
segment_density(
  x,
  n,
  Ls = 1e+06,
  deny,
  type = c("CBS", "HMM"),
  plot_origin = TRUE,
  boxplot = FALSE
)
}
\arguments{
\item{x}{the input gene GRanges}

\item{n}{the number of states}

\item{Ls}{segment length}

\item{deny}{GRanges of deny region}

\item{type}{the type of segmentation}

\item{plot_origin}{plot the gene density of original gene Granges}

\item{boxplot}{boxplot of gene density for each state}
}
\description{
Segmentation based on gene density
}
