% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentDensity.R
\name{segmentDensity}
\alias{segmentDensity}
\title{Genomic segmentation based on gene density}
\usage{
segmentDensity(x, n, L_s = 1e+06, exclude = NULL, type = c("cbs", "hmm"))
}
\arguments{
\item{x}{the input gene GRanges}

\item{n}{the number of states}

\item{L_s}{segment length}

\item{exclude}{GRanges of excluded region}

\item{type}{the type of segmentation, either "cbs" (which will
use DNAcopy to segment) or "hmm" (which will use RcppHMM).
The packages are not imported by nullranges, but must be installed
by the user}
}
\value{
a GRanges with metadata columns containing:
\itemize{
\item{state} {segmentation state}
\item{counts} {average number of genes}
}
}
\description{
Genomic segmentation based on gene density
}
\examples{

n <- 10000
library(GenomicRanges)
gr <- GRanges("chr1", IRanges(round(
  c(runif(n/4,1,991), runif(n/4,1001,3991),
    runif(n/4,4001,4991), runif(n/4,7001,9991))),
  width=10), seqlengths=c(chr1=10000))
gr <- sort(gr)
exclude <- GRanges("chr1", IRanges(5001,6000), seqlengths=c(chr1=10000))
seg <- segmentDensity(gr, n=3, L_s=100, exclude=exclude, type="cbs")

}
