% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_density.R
\name{segmentDensity}
\alias{segmentDensity}
\title{Genomic segmentation based on gene density}
\usage{
segmentDensity(x, n, L_s = 1e+06, deny, type = c("cbs", "hmm"))
}
\arguments{
\item{x}{the input gene GRanges}

\item{n}{the number of states}

\item{L_s}{segment length}

\item{deny}{GRanges of deny region}

\item{type}{the type of segmentation, either "cbs" (which will
use DNAcopy to segment) or "hmm" (which will use RcppHMM).
The packages are not imported by nullranges, but must be installed
by the user}
}
\description{
Genomic segmentation based on gene density
}
